Fix for #71 - markers now work bidirectionally when clicked.
Refactored the code so that IDs are unique and geocoding is done as a separate step.
This was needed so that when developing, we could avoid hitting the city's website
unless necessary.

Updated Readme with information about how to get started with development.

Added .vscode/launch.json so that development in Visual Studio Code
is easier.

Added sorting operations before file writes so that art updates can be
more easily compared with version control.

Cleaned up files which are no longer used.